+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## [Unreleased]
 
 ### Added
+- Started moving ui labels to admin settings page.
+- Icons: basket, trash
+- Cart view route and component
+- Success View
+- Cart state when not empty
 - Header component with brys logo SVG, cat icon, and hamburger menu
 - Footer component with marta leshak SVG and cat illustration
 - Content component with product grid layout
@@ -17,3 +22,4 @@
 
 ### Fixed
 - R2 bucket type compatibility issue with Payload storage plugin
+- Ordering process.